split io files per element (instead of read.py and write.py)
