--- src/spatialdata/_core/__init__.py
@@ -1,1 +1,0 @@
-from spatialdata._core.spatialdata import SpatialData
