+++ src/spatialdata/_io/__init__.py
@@ -1,8 +1,14 @@
 from spatialdata._io.format import SpatialDataFormatV01
-from spatialdata._io.write import (
-    write_image,
-    write_labels,
-    write_points,
-    write_shapes,
-    write_table,
-)
+from spatialdata._io.io_points import write_points
+from spatialdata._io.io_raster import write_image, write_labels
+from spatialdata._io.io_shapes import write_shapes
+from spatialdata._io.io_table import write_table
+
+__all__ = [
+    "write_image",
+    "write_labels",
+    "write_points",
+    "write_shapes",
+    "write_table",
+    "SpatialDataFormatV01",
+]
