+++ src/spatialdata/_io/io_table.py
@@ -1,0 +1,26 @@
+import zarr
+from anndata import AnnData
+from anndata._io.specs import write_elem as write_adata
+from ome_zarr.format import Format
+
+from spatialdata._io.format import CurrentTablesFormat
+
+
+def write_table(
+    table: AnnData,
+    group: zarr.Group,
+    name: str,
+    group_type: str = "ngff:regions_table",
+    fmt: Format = CurrentTablesFormat(),
+) -> None:
+    region = table.uns["spatialdata_attrs"]["region"]
+    region_key = table.uns["spatialdata_attrs"].get("region_key", None)
+    instance_key = table.uns["spatialdata_attrs"].get("instance_key", None)
+    fmt.validate_table(table, region_key, instance_key)
+    write_adata(group, name, table)  # creates group[name]
+    tables_group = group[name]
+    tables_group.attrs["spatialdata-encoding-type"] = group_type
+    tables_group.attrs["region"] = region
+    tables_group.attrs["region_key"] = region_key
+    tables_group.attrs["instance_key"] = instance_key
+    tables_group.attrs["version"] = fmt.version
