+++ tests/_core/test_rasterize.py
@@ -4,7 +4,7 @@ from spatial_image import SpatialImage
 
 from spatialdata._core._rasterize import rasterize
 from spatialdata._core.core_utils import get_axis_names, get_spatial_axes
-from spatialdata._io.write import _iter_multiscale
+from spatialdata._io._utils import _iter_multiscale
 from tests.conftest import _get_images, _get_labels
